+++ src/client/tair_client_api.cpp
@@ -50,7 +50,7 @@ void tair_client_api::setup_cache(int area, size_t capacity) {
         cache_impl[area] = new data_entry_local_cache(capacity);
 }
 
-void tair_client_api::setup_cache(int area, size_t capacity, uint64_t expire_time) {
+void tair_client_api::setup_cache(int area, size_t capacity, int64_t expire_time) {
     if (area < 0 || area >= TAIR_MAX_AREA_COUNT)
         return;
     if (NULL == cache_impl[area]) {
@@ -118,7 +118,7 @@ void tair_client_api::close() {
 int tair_client_api::put(int area,
                          const data_entry &key,
                          const data_entry &data,
-                         int expire,
+                         int64_t expire,
                          int version,
                          bool fill_cache) {
     int ret = impl == NULL ? TAIR_RETURN_NOT_INIT : impl->put(area, key, data, expire, version, fill_cache);
@@ -148,7 +148,7 @@ int tair_client_api::mc_ops(int8_t mc_opcode,
                             int area,
                             const data_entry *key,
                             const data_entry *value,
-                            int expire,
+                            int64_t expire,
                             int version,
                             callback_mc_ops_pt callback, void *arg) {
     return impl == NULL ? TAIR_RETURN_NOT_INIT : impl->mc_ops(mc_opcode, area, key, value,
@@ -245,7 +245,7 @@ int tair_client_api::remove(int area,
 int tair_client_api::async_put(int area,
                                const data_entry &key,
                                const data_entry &data,
-                               int expire,
+                               int64_t expire,
                                int version,
                                void (*cb)(int ret, void *args),
                                void *args) {
@@ -311,7 +311,7 @@ int tair_client_api::prefix_gets(int area, const data_entry &pkey, const tair_da
 }
 
 int tair_client_api::prefix_put(int area, const data_entry &pkey, const data_entry &skey,
-                                const data_entry &value, int expire, int version) {
+                                const data_entry &value, int64_t expire, int version) {
     return impl == NULL ? TAIR_RETURN_NOT_INIT : impl->prefix_put(area, pkey, skey, value, expire, version);
 }
 
@@ -386,10 +386,10 @@ int tair_client_api::minvalid(int area,
 
 int tair_client_api::incr(int area,
                           const data_entry &key,
-                          int count,
-                          int *ret_count,
-                          int init_value/* = 0*/,
-                          int expire /*= 0*/) {
+                          int64_t count,
+                          int64_t *ret_count,
+                          int64_t init_value/* = 0*/,
+                          int64_t expire /*= 0*/) {
     if (area < 0 || count < 0 || expire < 0) {
         return TAIR_RETURN_INVALID_ARGUMENT;
     }
@@ -400,10 +400,10 @@ int tair_client_api::incr(int area,
 
 int tair_client_api::decr(int area,
                           const data_entry &key,
-                          int count,
-                          int *ret_count,
-                          int init_value/* = 0*/,
-                          int expire /*= 0*/) {
+                          int64_t count,
+                          int64_t *ret_count,
+                          int64_t init_value/* = 0*/,
+                          int64_t expire /*= 0*/) {
     if (area < 0 || count < 0 || expire < 0) {
         return TAIR_RETURN_INVALID_ARGUMENT;
     }
@@ -415,15 +415,15 @@ int tair_client_api::decr(int area,
 
 int tair_client_api::add_count(int area,
                                const data_entry &key,
-                               int count,
-                               int *ret_count,
-                               int init_value /*= 0*/) {
+                               int64_t count,
+                               int64_t *ret_count,
+                               int64_t init_value /*= 0*/) {
 
     return impl == NULL ? TAIR_RETURN_NOT_INIT : impl->add_count(area, key, count, ret_count, init_value);
 
 }
 
-int tair_client_api::set_count(int area, const data_entry &key, int count, int expire, int version) {
+int tair_client_api::set_count(int area, const data_entry &key, int64_t count, int64_t expire, int version) {
     return impl == NULL ? TAIR_RETURN_NOT_INIT : impl->set_count(area, key, count, expire, version);
 }
 
@@ -437,7 +437,7 @@ int tair_client_api::unlock(int area, const data_entry &key) {
 
 int tair_client_api::expire(int area,
                             const data_entry &key,
-                            int expire) {
+                            int64_t expire) {
     return impl == NULL ? TAIR_RETURN_NOT_INIT : impl->expire(area, key, expire);
 }
 //int tair_client_api::removeArea(int area)
